+++ packages/duck-ui/src/commands/template/template.libs.ts
@@ -6,14 +6,14 @@ import { get_registry_index, get_registry_item } from '~/utils/get-registry'
 import { registry_component_install } from '~/utils/registry-mutation'
 import { spinner as Spinner } from '~/utils/spinner'
 
-export async function template_command_action(args: string[], opt: templateop) {
-  const spinner = Spinner('initializing...').start()
-  try 
-
-
-    spinner.succeed('🧑 Done.!, enjoy mr duck!🦆')
-    process.exit(0)
-  } catch (error) {
-    spinner.fail('🦆 Something went wrong')
-  }
-}
+// export async function template_command_action(args: string[], opt: templateop) {
+//   const spinner = Spinner('initializing...').start()
+//   try
+//
+//
+//     spinner.succeed('🧑 Done.!, enjoy mr duck!🦆')
+//     process.exit(0)
+//   } catch (error) {
+//     spinner.fail('🦆 Something went wrong')
+//   }
+// }