+++ packages/duck-ui/src/commands/template/template.ts
@@ -1,6 +1,6 @@
 import { Command } from 'commander'
 import { template_command_config } from './template.constants'
-import { template_command_action } from './template.libs'
+// import { template_command_action } from './template.libs'
 
 const { name, description, options, argumentss } = template_command_config
 const { option_1, option_2 } = options
@@ -14,7 +14,7 @@ export function template_command(): Command {
     .argument(arg_1.name, arg_1.description, arg_1.defaultValue)
     .option(option_1.flags, option_1.description, option_1.defaultValue)
     .option(option_2.flags, option_2.description, option_2.defaultValue)
-    .action(template_command_action)
+  // .action(template_command_action)
 
   return template_command
 }