+++ packages/duck-ui/src/utils/registry-mutation/registry-mutation.lib.ts
@@ -233,6 +233,7 @@ export async function process_component_dependencies(
     spinner.text = `🔧 Installing ${highlighter.info(allDependencies.length)} dependencies...`
 
     const packageManager = await get_package_manager(process.cwd())
+    console.log(packageManager)
     const { failed: installation_step_1 } = await execa(
       packageManager,
       [packageManager !== 'npm' ? 'add' : 'install', 'lucide-react', ...allDependencies],