+++ packages/duck-motion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gentleduck/motion
 
+## 0.1.5
+
+### Patch Changes
+
+- enhanced the animations
+
 ## 0.1.4
 
 ### Patch Changes