+++ packages/duck-ui/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @gentleduck/duck-ui
 
+## 1.0.7
+
+### Patch Changes
+
+- fixed registry
+
+## 1.0.6
+
+### Patch Changes
+
+- fixed registry
+
+## 1.0.5
+
+### Patch Changes
+
+- new release
+
+## 1.0.4
+
+### Patch Changes
+
+- new registry
+
 ## 1.0.3
 
 ### Patch Changes